requirements.txt
numpy==1.26.0
scikit-learn>=1.4.0
PyYAML==6.0.1
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.15.0
streamlit>=1.53.0
pytest>=7.0.0
jinja2>=3.0.0
